Organized the project code into different folders and files
