+++ ask_away/lib/main.dart
@@ -1,12 +1,9 @@
-import 'package:ask_away/Menu.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:ask_away/Question.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:ask_away/Talk.dart';
 
-import 'UserProfile.dart';
+import 'components/MenuComponent.dart';
+import 'components/QuestionComponent.dart';
 
 String appTitle = 'Ask Away';
 
@@ -57,7 +54,6 @@ void createQuestion() {
   return;
 }
 
-
 class MyHomePageState extends StatefulWidget {
   @override
   _MyHomePageState createState() => new _MyHomePageState();
@@ -82,28 +78,4 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 }
 
-class MyTalksPage extends StatefulWidget {
-  @override
-  State<MyTalksPage> createState() {
-    return new MyTalksPageState();
-  }
-}
 
-class MyTalksPageState extends State<MyTalksPage> {
-  TalkList tList = new TalkList();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        backgroundColor: Colors.blue[50],
-        appBar: AppBar(
-          title: Text('Talks'),
-        ),
-        body: new GestureDetector(
-            onTap: () {
-              FocusScope.of(context).requestFocus(new FocusNode());
-            },
-            child: tList),
-        drawer: MyDrawer());
-  }
-}
