+++ ask_away/lib/models/Vote.dart
@@ -1,13 +1,11 @@
 import 'User.dart';
 
-enum VoteType {
-  up,
-  down
-}
+enum VoteType { up, down }
 
 class Vote {
   User user;
   VoteType type;
+
   // Casted time
   Vote(this.type);
 }
